fixed alot of my response logic

+++ Car-Ali-App/src/pages/DashBoardPage/DashBoard/BidsBoard/BidLister/BidLister.js
@@ -3,15 +3,24 @@ import styles from "./BidLister.module.css";
 import BidItem from "./BidItem/BidItem";
 
 const bidLister = props => {
-	const bidItems = props.bidItems.map(bid => (
-		<BidItem
-			key={bid.id}
-			listingId={bid.listing_id}
-			cost={bid.bid}
-			bidId={bid.id}
-			deleteItem={props.deleteItem}
-		/>
-	));
+
+	let bidItems = null;
+	
+	if(props.bidItems.length){
+		bidItems = props.bidItems.map(bid => (
+			<BidItem
+				key={bid.id}
+				listingId={bid.listing_id}
+				cost={bid.bid}
+				bidId={bid.id}
+				deleteItem={props.deleteItem}
+			/>
+		));
+	}
+
+
+	
+	
 
 	return <div className={styles.BidLister}>{bidItems}</div>;
 };
